Adding the action.yaml file (pair w/ dotnet.yaml)
I added the action.yaml to go with the workflow I created for the GitHub action I am working on for when I get this project "customer" ready. This will put the necessary build dependencies into the GitHub build process to make it slightly easier on any clients who do not understand what to do with more technical aspects of how assemblies work in the process of executing my code.